@@ -89,8 +89,7 @@ class MainDrawer extends StatelessWidget {
                                 color: AppColors.primaryLight,
                                 borderRadius: BorderRadius.circular(20),
                                 border: Border.all(
-                                  color: AppColors.primary.withOpacity(0.1,
-                                  ),
+                                  color: AppColors.primary.withValues(alpha: 0.1),
                                   width: 2,
                                 ),
                               ),
@@ -299,7 +298,7 @@ class MainDrawer extends StatelessWidget {
                         padding: const EdgeInsets.symmetric(vertical: 16),
                         decoration: BoxDecoration(
                           border: Border.all(
-                            color: AppColors.error.withOpacity(0.2),
+                            color: AppColors.error.withValues(alpha: 0.2),
                           ),
                           borderRadius: BorderRadius.circular(16),
                         ),
